apps/worker/requirements.txt
pydantic>=2
python-dotenv>=1
httpx>=0.27
tenacity>=8
orjson>=3
eth-abi>=5
pydantic-settings>=2
# uvloop: platform-specific; install manually on Linux/macOS for better async performance
uvloop
